+++ src/pro/Carousel/CarouselComp.tsx
@@ -5,7 +5,7 @@ import { ArrowLeft, ArrowRight } from 'lucide-react';
 import * as React from 'react';
 
 import { cn } from '@/lib/utils';
-import { ButtonComp } from '../Button';
+import { ButtonComp } from '../../components/Button';
 
 type CarouselApi = UseEmblaCarouselType[1];
 type UseCarouselParameters = Parameters<typeof useEmblaCarousel>;