+++ src/pro/index.ts
@@ -1,0 +1,155 @@
+/**
+ * @dashflowx/core-pro entry (C04/C05).
+ * Not imported by the free package. Peer-depends on @dashflowx/core.
+ */
+export { Calendar } from './Calendar';
+
+export {
+  Carousel,
+  CarouselComp,
+  CarouselContent,
+  CarouselItem,
+  CarouselNext,
+  CarouselPrevious,
+  type CarouselApi,
+} from './Carousel';
+
+export {
+  Collapsible,
+  CollapsibleComp,
+  CollapsibleContent,
+  CollapsibleTrigger,
+} from './Collapsible';
+
+export {
+  Command,
+  CommandComp,
+  CommandDialog,
+  CommandEmpty,
+  CommandGroup,
+  CommandInput,
+  CommandItem,
+  CommandList,
+  CommandSeparator,
+  CommandShortcut,
+} from './Command';
+
+export {
+  ContextMenu,
+  SimpleContextMenu,
+} from './ContextMenu';
+
+export type { ContextMenuItemType } from './ContextMenu/utils';
+export {
+  parseArrayProp as parseContextMenuArrayProp,
+  getVariantClasses as getContextMenuVariantClasses,
+  getThemeClasses as getContextMenuThemeClasses,
+  getSubMenuClasses as getContextMenuSubMenuClasses,
+  isItemDisabled as isContextMenuItemDisabled,
+  createMenuItem as createContextMenuItem,
+  createLabelItem as createContextMenuLabelItem,
+  createSeparatorItem as createContextMenuSeparatorItem,
+  createSubMenuItem as createContextMenuSubMenuItem,
+  createRadioGroup as createContextMenuRadioGroup,
+  createCheckboxItem as createContextMenuCheckboxItem,
+} from './ContextMenu/utils';
+
+export type { DropdownMenuItemType } from './DropDownMenu/utils';
+export {
+  parseArrayProp,
+  getVariantClasses,
+  getThemeClasses,
+  getSubMenuClasses,
+  getBackgroundClasses,
+  isItemDisabled,
+  createMenuItem,
+  createLabelItem,
+  createSeparatorItem,
+  createGroupItem,
+  createSubMenuItem,
+} from './DropDownMenu/utils';
+
+export { DatePicker } from './DatePicker';
+
+export {
+  DropdownMenu,
+  DropdownMenuCheckboxItem,
+  DropdownMenuComp,
+  DropdownMenuContent,
+  DropdownMenuGroup,
+  DropdownMenuItem,
+  DropdownMenuLabel,
+  DropdownMenuPortal,
+  DropdownMenuRadioGroup,
+  DropdownMenuRadioItem,
+  DropdownMenuSeparator,
+  DropdownMenuShortcut,
+  DropdownMenuSub,
+  DropdownMenuSubContent,
+  DropdownMenuSubTrigger,
+  DropdownMenuTrigger,
+} from './DropDownMenu';
+
+export {
+  HoverCard,
+  HoverCardComp,
+  HoverCardContent,
+  HoverCardTrigger,
+} from './HoverCard';
+
+export {
+  InputOTP,
+  InputOTPGroup,
+  InputOTPSeparator,
+  InputOTPSlot,
+  InputOtpComp,
+} from './InputOTP';
+
+export {
+  Menubar,
+  DynamicMenubar,
+  MenubarCheckboxItem,
+  MenubarComp,
+  MenubarContent,
+  MenubarGroup,
+  MenubarItem,
+  MenubarLabel,
+  MenubarMenu,
+  MenubarPortal,
+  MenubarRadioGroup,
+  MenubarRadioItem,
+  MenubarSeparator,
+  MenubarShortcut,
+  MenubarSub,
+  MenubarSubContent,
+  MenubarSubTrigger,
+  MenubarTrigger,
+} from './Menubar';
+
+export type {
+  MenubarSubItem,
+  MenubarMenuItem,
+  MenubarMenuConfig,
+  DynamicMenubarProps,
+} from './Menubar';
+
+export { MenuList, MenuListComp, MenuListOne, type iDfxMenu } from './MenuList';
+
+export {
+  NavigationMenu,
+  NavigationMenuComp,
+  NavigationMenuContent,
+  NavigationMenuIndicator,
+  NavigationMenuItem,
+  NavigationMenuLink,
+  NavigationMenuList,
+  NavigationMenuTrigger,
+  NavigationMenuViewport,
+} from './NavigationMenu';
+
+export {
+  Resizable,
+  ResizableHandle,
+  ResizablePanel,
+  ResizablePanelGroup,
+} from './Resizable';