+++ vite.config.ts
@@ -26,5 +26,5 @@ export default defineConfig({
       },
     },
   },
-  plugins: [react(), dts({ rollupTypes: true })],
+  plugins: [react(), dts({ rollupTypes: true, skipDiagnostics: true })],
 });